refactor(follow-list): update page layout
