+++ packages/mesh/app/profile/member/[customId]/follower/page.tsx
@@ -1,9 +1,9 @@
+import EmptyFollowStatus from '@/app/profile/_components/empty-follow-status'
+import FollowerListItem from '@/app/profile/_components/following-list-item'
 import { GetMemberFollowerListDocument } from '@/graphql/__generated__/graphql'
 import fetchGraphQL from '@/utils/fetch-graphql'
 
 import { PageProps } from '../page'
-import EmptyFollowingStatus from './_components/empty-following-status'
-import FollowerListItem from './_components/follower-list-item'
 
 const FollowerPage = async ({ params, searchParams }: PageProps) => {
   const takeCount = 20
@@ -14,7 +14,11 @@ const FollowerPage = async ({ params, searchParams }: PageProps) => {
   })
   const followList = response?.member?.follower
   if (!followList || !followList.length)
-    return <EmptyFollowingStatus isVisitor={isVisitor} />
+    return (
+      <EmptyFollowStatus
+        content={isVisitor ? '這個人還沒有粉絲' : '目前還沒有粉絲'}
+      />
+    )
   return (
     <div className="flex max-w-[1120px] grow flex-col items-center sm:gap-5  sm:p-5 md:px-[70px] md:py-10 xl:w-maxMain">
       <div className="w-full bg-white px-5 pb-3 pt-4">
