+++ packages/mesh/app/profile/member/[customId]/following/page.tsx
@@ -1,11 +1,11 @@
+import EmptyFollowStatus from '@/app/profile/_components/empty-follow-status'
 import {
   GetMemberFollowingListDocument,
   GetMemberFollowingListQuery,
 } from '@/graphql/__generated__/graphql'
 import fetchGraphQL from '@/utils/fetch-graphql'
 
-import { PageProps } from '../page'
-import EmptyFollowingStatus from './_components/empty-following-status'
+import { type PageProps } from '../page'
 import FollowingList from './_components/following-list'
 
 export type FollowingListType = NonNullable<
@@ -25,8 +25,6 @@ const FollowingPage = async ({ params, searchParams }: PageProps) => {
   const isVisitor = params.customId !== searchParams.user
   const followPublisherResponse = response?.member?.follow_publisher || []
   const followResponse = response?.member?.following || []
-  const hasPublisherData = !!followPublisherResponse.length
-  const hasFollowingData = !!followResponse.length
   const followPublisherData = followPublisherResponse.map((followItem) => {
     return {
       ...followItem,
@@ -34,12 +32,20 @@ const FollowingPage = async ({ params, searchParams }: PageProps) => {
       name: followItem.title,
     }
   })
+  const hasPublisherData = !!followPublisherResponse.length
+  const hasFollowingData = !!followResponse.length
+
   if (!hasPublisherData && !hasFollowingData) {
-    return <EmptyFollowingStatus isVisitor={isVisitor} />
+    return (
+      <EmptyFollowStatus
+        content={
+          isVisitor ? '這個人還沒有追蹤中的對象' : '目前還沒有追蹤中的對象'
+        }
+      />
+    )
   }
   return (
-    <div className="flex max-w-[1120px] grow flex-col items-center sm:gap-5  sm:p-5 md:px-[70px] md:py-10 xl:w-maxMain">
-      {/**TODO: map list*/}
+    <div className="flex max-w-[1120px] grow flex-col items-center sm:gap-5 sm:p-5 md:px-[70px] md:py-10 xl:w-maxMain">
       <FollowingList
         title="媒體"
         followingList={followPublisherData as FollowingListType}
