+++ packages/mesh/app/profile/publisher/[publisherId]/follower/page.tsx
@@ -1,9 +1,9 @@
+import EmptyFollowStatus from '@/app/profile/_components/empty-follow-status'
+import FollowerListItem from '@/app/profile/_components/following-list-item'
 import { GetPublisherFollowerListDocument } from '@/graphql/__generated__/graphql'
 import fetchGraphQL from '@/utils/fetch-graphql'
 
 import { type PageProps } from '../page'
-import EmptyFollowingStatus from './_components/empty-following-status'
-import FollowerListItem from './_components/follower-list-item'
 
 const FollowerPage = async ({ params }: PageProps) => {
   const takeCount = 20
@@ -13,7 +13,7 @@ const FollowerPage = async ({ params }: PageProps) => {
   })
   const targetPublisher = response?.publishers && response?.publishers[0]
   const followList = targetPublisher?.follower || []
-  if (!followList.length) return <EmptyFollowingStatus />
+  if (!followList.length) return <EmptyFollowStatus content="目前還沒有粉絲" />
   return (
     <div className="flex max-w-[1120px] grow flex-col items-center sm:gap-5  sm:p-5 md:px-[70px] md:py-10 xl:w-maxMain">
       <div className="w-full bg-white px-5 pb-3 pt-4">
@@ -24,7 +24,6 @@ const FollowerPage = async ({ params }: PageProps) => {
                 key={customId}
                 followerId={id}
                 followerCustomId={customId || ''}
-                // TODO: Avatar change to shared components
                 followerAvatar={avatar || ''}
                 followerName={name || ''}
               />
